Refactor GameModel class hierarchy
Cover light sensor minigame (optional) #64
